refactor(website): enhance performance of chain retrieval process
