+++ packages/website/src/features/Search/PackageCard/Chain.tsx
@@ -1,7 +1,6 @@
-import { find, merge } from 'lodash';
 import { FC, useMemo } from 'react';
 import { Flex, Box, Text } from '@chakra-ui/react';
-import chains from '@/helpers/chains';
+import { chainsById } from '@/helpers/chains';
 import { Image } from '@chakra-ui/react';
 
 type ChainData = {
@@ -11,62 +10,13 @@ type ChainData = {
   [key: string]: any; // This allows for additional properties without having to specify each one.
 };
 
-const metadata = {
-  arbitrum: {
-    color: '#96bedc',
-  },
-  avalanche: {
-    color: '#e84141',
-  },
-  base: {
-    color: '#0052ff',
-  },
-  bsc: {
-    color: '#ebac0e',
-  },
-  cronos: {
-    color: '#002D74',
-  },
-  mainnet: {
-    color: '#37367b',
-  },
-  hardhat: {
-    color: '#f9f7ec',
-  },
-  optimism: {
-    color: '#ff5a57',
-  },
-  polygon: {
-    color: '#9f71ec',
-  },
-  zora: {
-    color: '#000000',
-  },
-  scroll: {
-    color: '#ffeeda',
-  },
-  gnosis: {
-    color: '#3e6957',
-  },
-};
-
-const newChain = {
-  id: 13370,
-  name: 'Cannon',
-  color: 'gray.400',
-};
-
 // Also see https://github.com/rainbow-me/rainbowkit/blob/7b31af2431cda62bf215a84b9d40fb7f66f24381/packages/rainbowkit/src/components/RainbowKitProvider/provideRainbowKitChains.ts
 const Chain: FC<{
   id: number;
   isSmall?: boolean;
+  chainData?: ChainData;
 }> = ({ id, isSmall }) => {
-  const chain = useMemo(() => {
-    const enrichedChainData: Record<string, ChainData> = {
-      ...merge({ cannon: newChain }, chains, metadata),
-    };
-    return find(enrichedChainData, (chain: ChainData) => chain.id === id);
-  }, [id]);
+  const chain = useMemo(() => chainsById[id], [id]);
 
   const name = chain?.name || 'Unknown Chain';
   const color = chain?.color || 'gray.600';
